FR-1 fix rename question reducer
