+++ src/store/duck/question/index.ts
@@ -1,2 +1,2 @@
-export * from './question';
+export * from './questionReducer';
 export * from './actions';
